refactor(captcha): split captcha interfaces into separate files
Split the single captcha.interface.ts file into separate files for each interface:
- ICaptchaChallenge
- ICaptchaValidationRequest
- ICaptchaValidationResult

This improves code organization and makes each interface more discoverable. Also updated
imports in related files and removed commented code from captcha-type.enum.ts.
