+++ src/domain/enum/captcha-type.enum.ts
@@ -3,7 +3,4 @@
  */
 export enum ECaptchaType {
 	CLICK = "click",
-	// Future types can be added here
-	// SLIDER = "slider",
-	// IMAGE = "image",
 }
