+++ src/domain/interface/captcha-challenge.interface.ts
@@ -1,0 +1,9 @@
+/**
+ * Interface representing captcha challenge data
+ */
+export interface ICaptchaChallenge {
+	createdAt: Date;
+	expiresAt: Date;
+	id: string;
+	type: string;
+}
