+++ src/domain/interface/captcha-validation-request.interface.ts
@@ -1,0 +1,7 @@
+/**
+ * Interface representing captcha validation request
+ */
+export interface ICaptchaValidationRequest {
+	challengeId: string;
+	response: boolean | string;
+}
