+++ src/domain/interface/captcha-validation-result.interface.ts
@@ -1,0 +1,8 @@
+/**
+ * Interface representing captcha validation result
+ */
+export interface ICaptchaValidationResult {
+	error?: string;
+	isSuccess: boolean;
+	token?: string;
+}
