+++ src/domain/interface/index.ts
@@ -1,2 +1,4 @@
+export type * from "./captcha-challenge.interface";
 export type * from "./captcha-service.interface";
-export type * from "./captcha.interface";
+export type * from "./captcha-validation-request.interface";
+export type * from "./captcha-validation-result.interface";
